some changes

+++ group2 pages/submit/submit_edit.php
@@ -6,32 +6,34 @@ $id = isset($_POST['id']) ? $_POST['id'] : '';
 $name = isset($_POST['product_name']) ? $_POST['product_name'] : '';
 $price = isset($_POST['product_price']) ? $_POST['product_price'] : '';
 $category = isset($_POST['category']) ? $_POST['category'] : '';
-$image = isset($_FILES['image']['name']) ? $_FILES['image']['name'] : $_POST['pimage'];
+$status = isset($_POST['status']) ? $_POST['status'] : 'out_of_stock';
 
-var_dump($_FILES['image']['name']);
-var_dump($_POST['pimage']);
 
-if (isset($_FILES["image"]["name"])) {
+
+if (!empty($_FILES['image']['name'])) {
+    $image = $_FILES['image']['name'];
     $image_ext = explode(".", $image);
     $image_ext = end($image_ext);
     $image_ext = strtolower($image_ext);
-    if (in_array($image_ext, array('jpg', 'jpeg', 'png'))) {
+    if (in_array($image_ext, array('jpg', 'jpeg', 'png'))&&$_FILES["image"]["size"]<25000000) {
         $file_size = $_FILES["image"]["size"];
-        $file_type = $_FILES["image"]["type"];
         $file_tmp = $_FILES["image"]["tmp_name"];
         $image_name ="../uploads/". time() . "_" . $image;
         $moved = move_uploaded_file($file_tmp, "../$image_name");
         if (!$moved) {
             $image_name = null;
         }
-        edit_product($id,$name,$price,$category,$image_name);
+        edit_product($id,$name,$price,$category,$image_name,$status);
         unlink("../".$_POST['pimage']);
-    }
+        }
+
 }else{
-    edit_product($id,$name,$price,$category,$image);
+    $image = $_POST['pimage'];
+    edit_product($id,$name,$price,$category,$image,$status);
+
 }
 
-
+var_dump($_FILES["image"]["size"]);
 header('location:../all_products.php');
